Change some of the project gql fields to use cache_resolve
